simulation changes
- move simulation structs to correct locations
- add helper functions
- remove code that prevented specific sim updates that didn't exist in halo 3
- move simulation_player_joined_game and simulation_player_left_game to simulation.cpp
