+++ xlive/Blam/Engine/simulation/simulation.h
@@ -1,58 +1,17 @@
 #pragma once
-#include "simulation_encoding.h"
-#include "simulation_type_collection.h"
-#include "simulation_watcher.h"
-#include "simulation_world.h"
 
-#include "game/player_control.h"
-#include "units/unit_control.h"
+/* prototypes */
 
-enum
-{
-	k_maximum_simulation_player_updates = 64
-};
+void simulation_apply_patches(void);
 
-struct simulation_update
-{
-	int32 update_number;
-	bool simulation_in_progress;
-	uint32 player_action_mask;
-	int32 field_C;
-	player_action player_actions[k_maximum_players];
-	uint32 unit_control_mask;
-	datum control_unit_index[k_maximum_players];
-	unit_control_data unit_control[k_maximum_players];
-	bool machine_update_valid;
-	simulation_machine_update machine_update;
-	int32 player_update_count;
-	simulation_player_update player_updates[k_maximum_simulation_player_updates];
-	bool flush_gamestate;
-	int32 verify_game_time;
-	uint32 verify_random_seed;
-};
-ASSERT_STRUCT_SIZE(struct simulation_update, 0x3BD8);
+void __cdecl simulation_player_joined_game(int32 player_index);
 
-struct s_simulation_globals
-{
-	bool initialized;
-	bool simulation_fatal_error;
-	bool engine_paused;
-	int32 field_4;
-	bool simulation_invalidate;
-	bool simulation_reset_pending;
-	bool simulation_reset_in_progress;
-	bool loading_saved_game;
-	c_simulation_world* world;
-	c_simulation_watcher* watcher;
-	c_simulation_type_collection* type_collection;
-};
-ASSERT_STRUCT_SIZE(s_simulation_globals, 24);
+void __cdecl simulation_player_left_game(int32 player_index);
 
-c_simulation_world* simulation_get_world();
-s_simulation_globals* simulation_get_globals();
+class c_simulation_world* simulation_get_world();
 
 void simulation_reset(void);
-bool simulation_reset_in_progress();
+bool simulation_reset_in_progress(void);
 
 void __cdecl simulation_update(void);
 
@@ -67,7 +26,7 @@ void simulation_notify_going_active(void);
 bool simulation_in_progress(void);
 void simulation_destroy_update(struct simulation_update* update);
 bool simulation_query_object_is_predicted(datum object_datum);
-c_simulation_type_collection* simulation_get_type_collection();
+class c_simulation_type_collection* simulation_get_type_collection();
 
 void __cdecl simulation_apply_before_game(const struct simulation_update* update);
 
@@ -79,8 +38,8 @@ void __cdecl simulation_update_aftermath(const struct simulation_update* update)
 
 void simulation_update_pregame(void);
 
-void __cdecl simulation_process_input(uint32 player_action_mask, const player_action* player_actions);
+void __cdecl simulation_process_input(uint32 player_action_mask, struct player_action const* player_actions);
 
 bool __cdecl simulation_get_machine_active_in_game(struct s_machine_identifier* machine_identifier);
 
-void simulation_apply_patches(void);
+void __cdecl simulation_build_player_updates(int32* player_update_count, int32 maximum_player_update_count, struct simulation_player_update* player_updates);
