+++ xlive/Blam/Engine/simulation/simulation_encoding.h
@@ -1,14 +1,4 @@
 #pragma once
-#include "game/players.h"
-#include "networking/network_constants.h"
-
-/* structures */
-
-struct simulation_machine_update
-{
-	uint32 machine_valid_mask;
-	s_machine_identifier identifiers[k_network_maximum_machines_per_session];
-};
 
 /* prototypes */
