+++ xlive/Blam/Engine/simulation/simulation_queue_entities.cpp
@@ -8,6 +8,7 @@
 #include "networking/network_event.h"
 #include "simulation/simulation.h"
 #include "simulation/simulation_entity_database.h"
+#include "simulation/simulation_world.h"
 
 // We get the absolute entity_def index from here as well as h2 needs it in order to create the game entity_def
 bool encode_simulation_queue_creation_to_buffer(uint8* out_buffer, int32 out_buffer_size, datum gamestate_index, s_simulation_queue_entity_data* data, uint32 initial_update_mask, int32* out_written_size);
