+++ xlive/Blam/Engine/simulation/simulation_world.cpp
@@ -161,10 +161,7 @@ void c_simulation_world::apply_simulation_queue(const c_simulation_queue* simula
 				simulation_queue_player_event_apply(element);
 				break;
 			case _simulation_queue_element_type_player_update_event:
-				if (!simulation_get_globals()->simulation_fatal_error)
-				{
-					simulation_queue_player_update_apply(element);
-				}
+				simulation_queue_player_update_apply(element);
 				break;
 			case _simulation_queue_element_type_gamestates_clear:
 				break;
