+++ xlive/Blam/Engine/text/text_group.cpp
@@ -1,8 +1,9 @@
 #include "stdafx.h"
 #include "text_group.h"
+
 #include "cache/cache_files.h"
-#include "main/game_preferences.h"
 #include "game/game_globals.h"
+#include "main/game_preferences.h"
 
 /* structures */
