+++ xlive/Project_Cartographer.vcxproj
@@ -3010,6 +3010,7 @@
     <ClInclude Include="Blam\Engine\simulation\simulation_queue_events.h" />
     <ClInclude Include="Blam\Engine\simulation\simulation_queue_global_events.h" />
     <ClInclude Include="Blam\Engine\simulation\simulation_type_collection.h" />
+    <ClInclude Include="Blam\Engine\simulation\simulation_update.h" />
     <ClInclude Include="Blam\Engine\simulation\simulation_view.h" />
     <ClInclude Include="Blam\Engine\simulation\simulation_view_telemetry.h" />
     <ClInclude Include="Blam\Engine\simulation\simulation_watcher.h" />
